requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.3
pillow==10.4.0
torchvision==0.19.1
matplotlib==3.9.2
torch==2.4.1 
requests==2.32.3